Refactor global message IDs to use separate message map
This seems to be required in order to be able to import messageIds with ts-node,
which is something we want to do to create a tool that generates message YAML
files from the defined IDs.

The old structure, where global messages were added in makeMessages() was
creating a circular import dependency, which normally is fine, but for some
reason does not work with ts-node.