Use '' if not editing and placeholder undefined
A value passed to controlled components should not be undefined or null.

https://react.dev/reference/react-dom/components/input
 #controlling-an-input-with-a-state-variable

Our test 'does trigger onChange with no text if explicitly allowed' for
ZUIEditTextInPlace found this but only printed a warning about it.

I don't think we ever use ZUIEditTextInPlace with allowEmpty without
also passing placeholder. Then the bug is not exposed anywhere.