Paste text instead of typing it in test
The tests '<TagDialog /> creates a basic tag' and '<TagDialog /> When
creating a new group, sends the new group properties to the onSubmit
callback instead of groupId' are consistently slow enough on my laptop
to trip Jest's default timeout of 5 seconds.

If I run the tests serially, it more or less narrowly fits into
5 seconds but by default, jest runs tests in parallell and then
execution is always too slow for at least one of these tests.

Profiling 'When creating a new group, sends the new group properties
to the onSubmit callback instead of groupId', I found that a lot of
time is spent typing the characters into the fields. While it is more
realistic to type the text into the fields, _how_ text is entered does
not seem to be what the authors were concerned about.

node --cpu-prof node_modules/jest-cli/bin/jest.js
    -i
    --testPathPattern TagDialog
    --testNamePattern 'When creating a new group, \
    sends the new group properties'

With paste instead of keyboard, execution time is approximately halved
on my potato machine. It won't make a significant impact on suite time,
but it makes a big difference for me, in that the suite succeeds. =)