Fix editor.js from hijacking focus when typing '/' outside of the editor
The problem arises from the editor.js having an event listener for pressing '/'
even when outside of the editor. This in combination with the editor not
unselecting the edited block when the editor is blurred results in a behaviour
where typing '/' anywhere else on the page resets focus back into the editor.js
window. See linked code [1] in the editor.js-repo.
This commit fixes the issue by using `stopImmediatePropagation` on the event
when the editor.js editor is mounted, but the event is triggered outside of the
editor.js editor, so that the code in [1] doesn't run. This can have some
side-effects, as all other subsequent keydown event handlers on `document` will
also be prevented. Luckily, we only seem to have one of these currently, and this
commit also makes sure that event handle fires first so that it is not prevented.

[1]: https://github.com/codex-team/editor.js/blob/58a7a9eeeadbfaef3c256576c8f37d64cb5d0edf/src/components/modules/ui.ts#L510-L513