Fix failing to redirect from `people/list/` due to missing try-catch
Fix a subtle bug due to an incorrect control flow. The API request
would throw an error if a user lacked the correct permission to access
a people list, but the code assumed that the API would return a falsy
value instead. The fix was simply to replace the `if-else` based control
flow with a `try-catch` dito, properly redirecting a user which performs
an unauthorized get request for some specific people list.