fix: load avatars, while also allowing for filtering, in LocalPerson
Before this commit, avatars would not load properly in LocalPerson, because when
we added the valueGetter to make filtering work, the properties that the other
functions relied on disappeared. In an ideal world, this would be caught by the
Typescript types, but I wasn't able to get them to work. Instead, this commit
alters the valueGetter so that it keeps all existing properties, and adds a
toString-method, which works with the filter functionality.